Move screenshot related classes to a new package
